Initial Commit
Two identical trains with a random distance apart from each other are placed on a single track that expands infinitely from east to west. The trains do not know where they are, but they do place a mark on the spot where they are initially placed, and each train is controlled by a script. The trains’ script can only understand 4 commands: move left, move right, go to line #, and if next to starting point go to line #. Assume that each command line takes an identical time to execute and each movement command moves the train by an equal distance.
Write a command script that when run on both trains at the beginning of their placement on the track, causes them to eventually but assuredly collide with each other.